Added modal component to get a grid of all the pictures the user can choose from for their profile picture.
