+++ src/pages/index.astro
@@ -7,7 +7,6 @@ import CardWithInput from '../components/CardWithInput.astro';
 // Importing database tools.
 import { app } from "../firebase/server";
 import { getAuth } from "firebase-admin/auth";
-import { getFirestore } from 'firebase-admin/firestore';
 
 // Importing user interface and user functions.
 import type { userModel } from '../pages/api/interfaces/user';
