+++ src/pages/signin.astro
@@ -66,8 +66,6 @@ if (Astro.cookies.has("session")) {
     getAuth,
     inMemoryPersistence,
     signInWithEmailAndPassword,
-    GoogleAuthProvider,
-    signInWithPopup,
   } from "firebase/auth";
   import { app } from "../firebase/client";
 
@@ -109,21 +107,4 @@ if (Astro.cookies.has("session")) {
       alert("Your password or email was incorrect, try again!")
     });
   });
-
-  // Code that enables google signin.
-  // const googleSignin = document.querySelector("#google") as HTMLButtonElement;
-  // googleSignin.addEventListener("click", async () => {
-  //   const provider = new GoogleAuthProvider();
-  //   const userCredential = await signInWithPopup(auth, provider);
-  //   const idToken = await userCredential.user.getIdToken();
-  //   const res = await fetch("/api/auth/signin", {
-  //     headers: {
-  //       Authorization: `Bearer ${idToken}`,
-  //     },
-  //   });
-
-  //   if (res.redirected) {
-  //     window.location.assign(res.url);
-  //   }
-  // });
 </script>
